u128 -> u64

+++ tvm_vm/src/executor/token.rs
@@ -23,7 +23,7 @@ pub(super) fn execute_ecc_mint(engine: &mut Engine) -> Status {
     fetch_stack(engine, 2)?;
     let x: u32 = engine.cmd.var(0).as_integer()?.into(0..=255)?;
     let y: VarUInteger32 =
-        VarUInteger32::from(engine.cmd.var(1).as_integer()?.into(0..=u128::MAX)?);
+        VarUInteger32::from(engine.cmd.var(1).as_integer()?.into(0..=u64::MAX)?);
     let mut data = ExtraCurrencyCollection::new();
     data.set(&x, &y)?;
     let mut cell = BuilderData::new();
